Switched all variables that were Double to BigDecimal

+++ src/main/java/com/ak98neon/currencyexchange/exchanger/entity/Commission.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "commissions")
@@ -14,10 +15,10 @@ import javax.persistence.Table;
 public class Commission {
     @EmbeddedId
     private CurrencyId currencyId;
-    private Double value;
+    private BigDecimal value;
 
     public Commission(final CurrencyId currencyId,
-                      final Double value) {
+                      final BigDecimal value) {
         this.currencyId = currencyId;
         this.value = value;
     }
